+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # NEXT RELEASE
 * Refactor `CommunicationCfg` and `EmailCfg` to their own files
+* `DocCfg.is_interesting` can be an `int` for the most interesting docs, replaces `DocCfg.is_very_interesting`
 
 # 1.10.0
 * Rename `DocCfg.description` to `note`